build: pin remark unwrap images to 4.0.1
I initially attempted to replace remark-unwrap-images with rehype-unwrap-images.
However, rehype-unwrap-images does not work as a drop in replacement.
rehype-unwrap-images does not remove the wrapper paragraph around images so
we will keep remark-unwrap-images for now.
I tried to migrate to the rehype plugin because the remark plugin is depcrecated in favor
of the rehype version.